third_party/blink/web_tests/fast/dom/custom/exception-from-constructor-expected.txt

CONSOLE WARNING: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests throwing an exception during registration of constructor

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.registerElement("x-a", {prototype: proto}) threw exception Exception thrown from getter.
PASS successfullyParsed is true

TEST COMPLETE
